Changed int to float in hw 1 task 1

+++ homework1/task_1.c
@@ -1,15 +1,15 @@
 #include <stdio.h>
 
-int f(int x)
+float f(float x)
 {
-    const int x2 = x * x;
+    const float x2 = x * x;
     return (x2 + 1) * (x2 + x) + 1;
 }
 
 int main()
 {
-    int x;
+    float x;
     printf("Enter x: ");
-    scanf("%d", &x);
-    printf("x^4 + x^3 + x^2 + x + 1 = %d\n", f(x));
+    scanf("%f", &x);
+    printf("x^4 + x^3 + x^2 + x + 1 = %f\n", f(x));
 }
